Added tests for react components. These tests are very basic, unfortunately I ran out of time to write more substantial tests.
